+++ NEWS.md
@@ -14,6 +14,7 @@
 - Reduce sampling of `rbmi` test to make tests shorter (#323)
 - Optimize shared tables in `test-tt_to_tblfile` (#323)
 - Changed forked `h_ancova` for the `tern` one
+- Add extra statistics to `a_eair100_j` and introduce scaling factor `num_p_year` (default = 100) (#361)
 - Removed ellipsis argument from `a_freq_resp_var_j` (#236) 
 - Removed `dplyr` from `junco` and replaced by base R #201
 - Replaced `assertthat` by `checkmate` for consistency #201